requirements.txt
aws-cdk-lib==2.38.1
constructs>=10.1.82
Flask==2.0.3
psycopg2-binary==2.9.3
gunicorn==20.1.0
PyMySQL==1.0.2




